# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MCP Server for coding workflows — file editing, code execution, git, package management, and more
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /files/read:
    get:
      operationId: readFile
      tags: [files]
      summary: Read file contents
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: encoding
          in: query
          required: false
          schema:
            type: string
            default: utf-8
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /execute:
    post:
      operationId: executeCode
      tags: [execution]
      summary: Execute code in a sandboxed environment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteRequest"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /terminal/sessions:
    get:
      operationId: listTerminalSessions
      tags: [terminal]
      summary: List terminal sessions
      responses:
        "200":
          description: Terminal sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TerminalSession"
    post:
      operationId: createTerminalSession
      tags: [terminal]
      summary: Create a new terminal session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminalSessionInput"
      responses:
        "201":
          description: Terminal session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalSession"
  /terminal/sessions/{sessionId}:
    delete:
      operationId: killTerminalSession
      tags: [terminal]
      summary: Kill a terminal session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session killed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /terminal/execute:
    post:
      operationId: terminalExecute
      tags: [terminal]
      summary: Execute a command in a terminal session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminalCommand"
      responses:
        "200":
          description: Command output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
        "403":
          description: Command blocked by safety controls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---- Git ----
...
  /packages/installed:
    get:
      operationId: listInstalledPackages
      tags: [packages]
      summary: List installed packages
      parameters:
        - name: manager
          in: query
          required: false
          schema:
            type: string
            default: npm
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: "."
      responses:
        "200":
          description: Package list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
  /packages/install:
    post:
      operationId: installPackage
      tags: [packages]
      summary: Install a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInstallRequest"
      responses:
        "200":
          description: Installation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageOperationResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /packages/uninstall:
    post:
      operationId: uninstallPackage
      tags: [packages]
      summary: Uninstall a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUninstallRequest"
      responses:
        "200":
          description: Uninstall result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageOperationResult"

  # ---- Environment Variables ----
...
  /mcp/tools/call:
    post:
      operationId: callMcpTool
      tags: [mcp]
      summary: Call an MCP tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McpToolCall"
      responses:
        "200":
          description: Tool call result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpToolResult"
        "404":
          description: Tool not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]

    # Files
...
    FileContent:
      type: object
      required: [path, content, size, encoding]
      properties:
        path:
          type: string
        content:
          type: string
        size:
          type: integer
        encoding:
          type: string
        mimeType:
          type: ["string", "null"]
...
    ExecuteRequest:
      type: object
      required: [language, code]
      properties:
        language:
          type: string
          enum: [node, python, bash]
        code:
          type: string
        timeout:
          type: integer
          default: 30
        workDir:
          type: ["string", "null"]
        env:
          type: ["object", "null"]
          additionalProperties:
            type: string
        args:
          type: array
          items:
            type: string
    ExecutionResult:
      type: object
      required: [stdout, stderr, exitCode, duration]
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        duration:
          type: number
        sessionId:
          type: ["string", "null"]
        timedOut:
          type: boolean
...
    TerminalSession:
      type: object
      required: [id, name, status, createdAt, shell]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, idle, terminated]
        createdAt:
          type: string
        shell:
          type: string
        workDir:
          type: ["string", "null"]
        pid:
          type: ["integer", "null"]
    TerminalSessionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        shell:
          type: string
          default: /bin/bash
        workDir:
          type: ["string", "null"]
        env:
          type: ["object", "null"]
          additionalProperties:
            type: string
    TerminalCommand:
      type: object
      required: [command]
      properties:
        command:
          type: string
        sessionId:
          type: ["string", "null"]
        timeout:
          type: integer
          default: 30
        workDir:
          type: ["string", "null"]

    # Git
...
    Package:
      type: object
      required: [name, version, manager]
      properties:
        name:
          type: string
        version:
          type: string
        manager:
          type: string
        description:
          type: ["string", "null"]
        isDev:
          type: boolean
    PackageInstallRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        version:
          type: ["string", "null"]
        manager:
          type: string
          default: npm
        dev:
          type: boolean
          default: false
        path:
          type: string
          default: "."
    PackageUninstallRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        manager:
          type: string
          default: npm
        path:
          type: string
          default: "."
    PackageOperationResult:
      type: object
      required: [success, output]
      properties:
        success:
          type: boolean
        output:
          type: string
        duration:
          type: number

    # Environment
...
    McpToolCall:
      type: object
      required: [name, arguments]
      properties:
        name:
          type: string
        arguments:
          type: object
          additionalProperties: {}
    McpToolResult:
      type: object
      required: [content, isError]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/McpContent"
        isError:
          type: boolean
    McpContent:
      type: object
      required: [type, text]
      properties:
        type:
          type: string
        text:
          type: string